Fix all platform-specific build warnings for warnings-as-errors
- Linux GCC: Add pragma to suppress false positive -Wmaybe-uninitialized on structured bindings
- Windows MSVC: Fix C4244 narrowing conversions with explicit casts
- Windows MSVC: Fix C4267 size_t to int conversions
- Windows MSVC: Suppress C4996 unsafe function warnings for gmtime/ctime
- Windows MSVC: Guard Clang-specific #pragma mark with #ifdef
- All platforms: Add static_cast for double/float conversions

+++ core/CalChartAnimationCompile.cpp
@@ -127,8 +127,8 @@ bool CompileState::Append(Command cmd)
     mBeatsRem -= numBeats;
 
     mWhichPos = End(cmd); // Move current point to new position
-    SetVarValue(Cont::Variable::DOF, MotionDirectionAtBeat(cmd, numBeats).getValue());
-    SetVarValue(Cont::Variable::DOH, FacingDirectionAtBeat(cmd, numBeats).getValue());
+    SetVarValue(Cont::Variable::DOF, static_cast<float>(MotionDirectionAtBeat(cmd, numBeats).getValue()));
+    SetVarValue(Cont::Variable::DOH, static_cast<float>(FacingDirectionAtBeat(cmd, numBeats).getValue()));
     mCmds.push_back(cmd);
     return true;
 }
